update to DARCDesignGenerator
Now we have a method enter_trial_design_and_response (in badapted repo) which automatically calls add_design_response_to_dataframe (which is user-defined) and it automatically takes care of incrementing the trial counter. So this is a bit more defensive and might avoid problems with user-code forgetting to increment the trial counter.